+++ etc/config.json
@@ -5,16 +5,6 @@
   "TIPdb": "线粒体数据库-fll-20190418.xlsx.MitoTIP-tRNA预测打分.json.db",
   "MTdisease": "线粒体数据库-fll-20190418.xlsx.疾病-位点库.json.db",
   "MTAFdb": "线粒体突变频率库-fll-20190522.xlsx.线粒体突变频率库-20190521.json.db",
-  "redisServer": "10.2.1.4:6380",
-  "resultColumn": [
-    "Gene Symbol",
-    "#Chr+Stop",
-    "MutationName",
-    "Zygosity",
-    "自动化判断",
-    "DiseaseName/ModeInheritance",
-    "IsACMG59"
-  ],
   "Acmg59Gene":"ACMG59Gene.list",
   "qualityColumn":[
     "样本编号",
